% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_mr_plots.R
\name{creating_mr_plots}
\alias{creating_mr_plots}
\title{Auto creating MR plots}
\usage{
creating_mr_plots(dat_harmonised, res, file_id)
}
\arguments{
\item{dat_harmonised}{harmonise后的数据}

\item{res}{初次MR分析后的数据}

\item{file_id}{file_id}
}
\description{
Auto creating MR plots
}
