% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immnune_online_reverse.R
\name{immnune_online_reverse}
\alias{immnune_online_reverse}
\title{Auto run 731 immune cells MR}
\usage{
immnune_online_reverse(
  exposure_data,
  output_path = "result_reverse.csv",
  multiprocess = FALSE,
  start_num = NA,
  end_num = NA,
  exposure_pval = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000
)
}
\arguments{
\item{exposure_data}{必须为数据框格式的暴露数据}

\item{output_path}{输出结果文件的名字，默认为result_reverse.csv，如果使用多线程分析则每个线程输出文件名字必须不同（否则会覆盖）}

\item{multiprocess}{是否启用多线程，默认关闭}

\item{start_num}{多线程的分析起点和终点，如果要使用多线程，则分别设置每个线程的起点和终点，比如线程1设置为1和350，线程2设置为351和731（总共731）}

\item{end_num}{多线程的分析起点和终点，如果要使用多线程，则分别设置每个线程的起点和终点，比如线程1设置为1和350，线程2设置为351和731（总共731）}

\item{exposure_pval}{筛选暴露的P值，默认为1e-5}

\item{clump_r2}{clump使用的r2和kb，默认为0.1和500}

\item{clump_kb}{clump使用的r2和kb，默认为0.1和500}
}
\description{
Auto run 731 immune cells MR
}
