% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_486.R
\name{met_486}
\alias{met_486}
\title{Auto run the 486 metabolites MR}
\usage{
met_486(
  exposure_path,
  exposure_name,
  outcome_path,
  outcome_name,
  p_threshold = 5e-08,
  kb = 10000,
  r2 = 0.001,
  outcome_source,
  result_path
)
}
\arguments{
\item{exposure_path}{暴露文件所在的目录，运行后会遍历该目录的所有文件，逐个与结局跑MR分析}

\item{outcome_path}{结局文件的路径（目前只支持单文件）}

\item{p_threshold}{筛选的P值，默认5e-8}

\item{kb}{Clump时的kb值，默认10000}

\item{r2}{Clump时的r2，默认0.001}

\item{outcome_source}{结局数据的来源，可选'ieu','finn'，会自动格式化结局}
}
\description{
Auto run the 486 metabolites MR
}
\author{
Jieyu
}
